backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
python-docx>=0.8.11
spacy>=3.1.0
nltk>=3.6.3
behave>=1.2.6
jinja2>=3.0.1
sqlalchemy>=1.4.23
pytest>=6.2.5
black>=21.7b0
flake8>=3.9.2
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4